+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/web/project/ProjectPlanningTimeController.java
@@ -15,21 +15,30 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.axelor.apps.project.web;
+package com.axelor.apps.hr.web.project;
 
-import com.axelor.apps.project.db.ProjectPlanning;
+import com.axelor.apps.hr.service.project.ProjectPlanningTimeService;
+import com.axelor.apps.project.db.ProjectPlanningTime;
+import com.axelor.exception.AxelorException;
 import com.axelor.i18n.I18n;
 import com.axelor.meta.schema.actions.ActionView;
 import com.axelor.meta.schema.actions.ActionView.ActionViewBuilder;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
 import com.axelor.rpc.Context;
+import com.google.inject.Inject;
 import com.google.inject.Singleton;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.Collection;
+import java.util.HashMap;
+import java.util.Locale;
 import java.util.Map;
 
 @Singleton
-public class ProjectPlanningController {
+public class ProjectPlanningTimeController {
+
+  @Inject private ProjectPlanningTimeService projectPlanningTimeService;
 
   public void showPlanning(ActionRequest request, ActionResponse response) {
 
@@ -50,16 +59,21 @@ public class ProjectPlanningController {
     }
 
     ActionViewBuilder builder =
-        ActionView.define(I18n.get("Project Planning")).model(ProjectPlanning.class.getName());
+        ActionView.define(I18n.get("Project Planning time"))
+            .model(ProjectPlanningTime.class.getName());
     String url = "project/planning";
-
-    if (!userIds.isEmpty()) {
-      url += "?userIds=" + userIds;
-      builder.domain("self.user.id in (:userIds)");
-      builder.context("userIds", userIds);
-    }
 
     builder.add("html", url);
     response.setView(builder.map());
+    response.setCanClose(true);
+  }
+
+  public void addMultipleProjectPlanningTime(ActionRequest request, ActionResponse response) throws AxelorException {
+
+    Context context = request.getContext();
+
+    projectPlanningTimeService.addMultipleProjectPlanningTime(context);
+
+    response.setCanClose(true);
   }
 }