+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Fill `qty` from `launchQty` in manuf order on filling prod process.
 - Declaration of exchanges of services and declaration of services
 - Project : project assignment on Sale order, Purchase order and Invoices.
+- PROJECT : Change project planning to project planning time with some extra fields.
 
 ## Improvements
 - Move project resource menu entry to project configuration menu.